% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{App server}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input}{app input}

\item{output}{app output}

\item{session}{app session}
}
\description{
App server
}
