% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{create_database}
\alias{create_database}
\title{Create database tables if they don't already exist}
\usage{
create_database(conn = db_conn())
}
\arguments{
\item{conn}{a connection to the database}
}
\value{
NULL, invisibly
}
\description{
Create database tables if they don't already exist
}
