% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{csl_to_vec}
\alias{csl_to_vec}
\title{Convert a comma-separated list to a vector}
\usage{
csl_to_vec(csl)
}
\arguments{
\item{csl}{a comma-separated list as a single-length character vector}
}
\value{
a multi-element character vector, split by comma
}
\description{
Convert a comma-separated list to a vector
}
