% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_games.R
\name{generate_game_name}
\alias{generate_game_name}
\title{Generate a random name for a game}
\usage{
generate_game_name(n = NULL)
}
\arguments{
\item{n}{optional number of players}
}
\description{
Generate a random name for a game
}
