% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humancluedo-package.R
\docType{package}
\name{humancluedo-package}
\alias{humancluedo}
\alias{humancluedo-package}
\title{humancluedo: Human Cluedo}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: 

Other contributors:
\itemize{
  \item Chris Brownlie \email{chris.brownlie@hotmail.co.uk} [contributor]
}

}
\keyword{internal}
