% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{launch_a_game}
\alias{launch_a_game}
\title{Show a game view in the app}
\usage{
launch_a_game(game_state, existing_cookie, game_id = NULL)
}
\arguments{
\item{game_state}{a GameState object to use to populate game information}

\item{game_id}{optional game ID. If left NULL, will use the currently active
game in the supplied game_state}
}
\description{
Show a game view in the app
}
