% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pkg_file_strict}
\alias{pkg_file_strict}
\title{Get a file from the human cluedo app}
\usage{
pkg_file_strict(...)
}
\arguments{
\item{...}{passed onto fs::path_package}
}
\description{
Get a file from the human cluedo app
}
