% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_sql.R
\name{set_player_id}
\alias{set_player_id}
\title{Set the player ID for a player in a game}
\usage{
set_player_id(id, player, game_id, conn)
}
\arguments{
\item{id}{the ID to add for the chosen player}

\item{player}{the name of the player within the chosen game}

\item{game_id}{the ID of the game for which a player ID is being added}

\item{conn}{a SQL connection}
}
\value{
the output of dbExecute
}
\description{
Set the ID used to identify players via their cookies
}
