% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{update_game_details}
\alias{update_game_details}
\title{Update basic game details like the name and the active player}
\usage{
update_game_details(state)
}
\description{
Update basic game details like the name and the active player
}
