% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_state.R
\name{GameState}
\alias{GameState}
\title{Custom class for holding state of current game}
\description{
Custom class for holding state of current game

Custom class for holding state of current game
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{conn}}{the database connection}

\item{\code{active_game}}{the ID of the game being represented}

\item{\code{active_player}}{the name of the currently active player,
in the active game}

\item{\code{player_status}}{a tibble indicating which players are alive}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GameState-initialise}{\code{GameState$initialise()}}
\item \href{#method-GameState-initialise_game}{\code{GameState$initialise_game()}}
\item \href{#method-GameState-clone}{\code{GameState$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GameState-initialise"></a>}}
\if{latex}{\out{\hypertarget{method-GameState-initialise}{}}}
\subsection{Method \code{initialise()}}{
Create a new game state
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameState$initialise(game_id = NULL, conn = db_conn())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{game_id}}{optional game ID}

\item{\code{conn}}{database connection object, defaults to using db_conn()}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a new object of class GameState
Populate game state using the database
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GameState-initialise_game"></a>}}
\if{latex}{\out{\hypertarget{method-GameState-initialise_game}{}}}
\subsection{Method \code{initialise_game()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameState$initialise_game(game_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{game_id}}{the ID of the game to use to populate the object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GameState-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GameState-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameState$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
